Move sample to proper package
