+++ gwt-polymer-demo/src/main/java/com/vaadin/components/gwt/polymer/client/sampler/Sampler.java
@@ -26,7 +26,6 @@ import com.vaadin.components.gwt.polymer.client.sampler.gwt.PaperTabsView;
 import com.vaadin.components.gwt.polymer.client.sampler.gwt.PaperTabsWidgetView;
 import com.vaadin.components.gwt.polymer.client.sampler.iron.IronCollapseSample;
 import com.vaadin.components.gwt.polymer.client.sampler.iron.IronSelectorSample;
-import com.vaadin.components.gwt.polymer.client.sampler.iron.RadioGroupSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.ButtonSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.CheckboxSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.DialogSample;
@@ -39,6 +38,7 @@ import com.vaadin.components.gwt.polymer.client.sampler.paper.MaterialSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.MenuSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.ProgressSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.RadioButtonSample;
+import com.vaadin.components.gwt.polymer.client.sampler.paper.RadioGroupSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.SpinnerSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.TabsSample;
 import com.vaadin.components.gwt.polymer.client.sampler.paper.ToastSample;
