+++ gwt-polymer-demo/src/main/java/com/vaadin/components/gwt/polymer/client/sampler/paper/RadioGroupSample.ui.xml
